# main_docs/openapi.yaml
# OpenAPI specification for the Telehealth API - to be updated as the API is developed
openapi: 3.0.3
info:
  title: Telehealth API (local)
  description: "OpenAPI spec for the Telehealth application (local/dev). Intended for import into tools/AI to reproduce server behavior."
  version: 1.0.0
servers:
  - url: http://localhost:3000
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required: [email, password]
    UploadRequest:
      type: object
      properties:
        fileName:
          type: string
        bucket:
          type: string
        metadata:
          type: object
    UploadResponse:
      type: object
      properties:
        signedUrl:
          type: string
        path:
          type: string
        token:
          type: string
    TranscribeRequest:
      type: object
      properties:
        audioPath:
          type: string
        audioUrl:
          type: string
        visitId:
          type: string
        metadata:
          type: object
      oneOf:
        - required: [audioPath]
        - required: [audioUrl]
    TranscribeResponse:
      type: object
      properties:
        jobId:
          type: string
        status:
          type: string
    Patient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        clinician_id:
          type: string
    Visit:
      type: object
      properties:
        id:
          type: string
        patient_id:
          type: string
        clinician_id:
          type: string
        started_at:
          type: string
          format: date-time
    NoteRequest:
      type: object
      properties:
        body:
          type: string
paths:
  /api/auth/signup:
    post:
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: Created
  /api/upload:
    post:
      tags: [Upload]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
            example:
              fileName: "sample.wav"
              bucket: "telehealth_audio"
      responses:
        '200':
          description: Signed URL created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
  /api/transcribe:
    post:
      tags: [Transcription]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TranscribeRequest'
            example:
              audioPath: "clinician/1/sample.wav"
              visitId: "visit-123"
      responses:
        '202':
          description: Job queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TranscribeResponse'
        '401':
          description: Unauthorized
        '413':
          description: Payload Too Large
  /api/patients:
    get:
      tags: [Patients]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of patients
    post:
      tags: [Patients]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Patient'
      responses:
        '201':
          description: Patient created
  /api/visits:
    get:
      tags: [Visits]
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of visits
    post:
      tags: [Visits]
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Visit'
      responses:
        '201':
          description: Visit created
  /api/visits/{id}/note:
    post:
      tags: [Visits]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NoteRequest'
      responses:
        '200':
          description: Note added
  /api/twilio/token:
    get:
      tags: [Video]
      security:
        - bearerAuth: []
      parameters:
        - name: room
          in: query
          schema:
            type: string
      responses:
        '200':
          description: { token }
  /api/patients/{id}/share:
    post:
      tags: [Patients]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                target_clinician_id:
                  type: string
      responses:
        '200':
          description: Share acknowledged